+++ src/lib.rs
@@ -190,7 +190,7 @@ impl App {
         let (diff_bases, diffs) = run_headless_sync_stage(&self.governor, || {
             let diff_bases = self
                 .repo
-                .resolve_diff_bases(&target, &bases, self.config.quiet);
+                .resolve_diff_bases(&self.config, &target, &bases);
             let diffs = self
                 .repo
                 .generate_diffs(&target, &diff_bases, self.config.quiet)?;
@@ -576,7 +576,7 @@ impl App {
             self.ensure_not_cancelled()?;
             let diff_bases = self
                 .repo
-                .resolve_diff_bases(&target, &bases, self.config.quiet);
+                .resolve_diff_bases(&self.config, &target, &bases);
             let diffs = self
                 .repo
                 .generate_diffs(&target, &diff_bases, self.config.quiet)?;